Don't filter countries when they're not supported from Locale.getISOCountries on older Android versions.
Fixes #6501
