+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/CountrySpec.kt
@@ -1,11 +1,11 @@
 package com.stripe.android.ui.core.elements
 
 import androidx.annotation.RestrictTo
+import com.stripe.android.core.model.CountryUtils
 import com.stripe.android.uicore.elements.CountryConfig
 import com.stripe.android.uicore.elements.CountryElement
 import com.stripe.android.uicore.elements.DropdownFieldController
 import com.stripe.android.uicore.elements.IdentifierSpec
-import com.stripe.android.uicore.elements.supportedBillingCountries
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
@@ -21,7 +21,7 @@ data class CountrySpec(
     override val apiPath: IdentifierSpec = IdentifierSpec.Country,
 
     @SerialName("allowed_country_codes")
-    val allowedCountryCodes: Set<String> = supportedBillingCountries
+    val allowedCountryCodes: Set<String> = CountryUtils.supportedBillingCountries
 ) : FormItemSpec() {
     fun transform(
         initialValues: Map<IdentifierSpec, String?>
