+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/forms/LinkCardForm.kt
@@ -1,16 +1,16 @@
 package com.stripe.android.ui.core.forms
 
 import androidx.annotation.RestrictTo
+import com.stripe.android.core.model.CountryUtils
 import com.stripe.android.ui.core.elements.CardBillingSpec
 import com.stripe.android.ui.core.elements.CardDetailsSectionSpec
 import com.stripe.android.ui.core.elements.LayoutSpec
 import com.stripe.android.uicore.elements.IdentifierSpec
-import com.stripe.android.uicore.elements.supportedBillingCountries
 
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 val LinkCardForm = LayoutSpec(
     listOf(
         CardDetailsSectionSpec(IdentifierSpec.Generic("card_details_section")),
-        CardBillingSpec(allowedCountryCodes = supportedBillingCountries)
+        CardBillingSpec(allowedCountryCodes = CountryUtils.supportedBillingCountries)
     )
 )
