+++ stripe-ui-core/src/test/java/com/stripe/android/uicore/elements/CountryConfigTest.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.uicore.elements
 
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.core.model.CountryUtils
 import com.stripe.android.uicore.R
 import org.junit.Test
 import java.util.Locale
@@ -62,7 +63,7 @@ class CountryConfigTest {
             locale = Locale.US
         ).displayItems
         val supportedCountries = CountryConfig(
-            onlyShowCountryCodes = supportedBillingCountries,
+            onlyShowCountryCodes = CountryUtils.supportedBillingCountries,
             locale = Locale.US
         ).displayItems
 
@@ -79,10 +80,10 @@ class CountryConfigTest {
 
         assertThat(
             defaultCountries.size
-        ).isEqualTo(249)
+        ).isEqualTo(235)
 
         assertThat(
             supportedCountries.size
-        ).isEqualTo(233)
+        ).isEqualTo(235)
     }
 }
